+++ lib/core/pages/timeline/timeline_card.dart
@@ -20,34 +20,49 @@ class EventCard extends StatelessWidget {
       decoration: BoxDecoration(
         color: AppPallete.timelineBg,
         borderRadius: BorderRadius.circular(12),
+        boxShadow: [
+          BoxShadow(
+            color: Colors.black.withValues(),
+            spreadRadius: 4,
+            blurRadius: 9,
+            offset: Offset(0, 4), // changes position of shadow
+          ),
+        ],
       ),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Center(
-            child: Text(title, style: TextStyle(
-              color: AppPallete.timelineGold,
-              fontSize: 30.33,
-              fontWeight: FontWeight.w600,
-              fontFamily: 'LemonMilkMedium',
-            ),),
-          ),
-          const SizedBox(height: 5),
-          Center(
-            child: Text(date, style: TextStyle(
-              color: AppPallete.redColor,
-              fontSize: 18.55,
-              fontWeight: FontWeight.w600,
-              fontFamily: 'LemonMilkMedium'),
+            child: Text(
+              title,
+              style: TextStyle(
+                color: AppPallete.timelineGold,
+                fontSize: 30.33,
+                fontWeight: FontWeight.w600,
+                fontFamily: 'LemonMilkMedium',
+              ),
             ),
           ),
           const SizedBox(height: 5),
-          Text(description, style: TextStyle(
-            color: Colors.white,
-            fontSize: 17.5,
-            fontWeight: FontWeight.w600,
-            fontFamily: 'LeagueSpartan',
+          Center(
+            child: Text(
+              date,
+              style: TextStyle(
+                  color: AppPallete.redColor,
+                  fontSize: 18.55,
+                  fontWeight: FontWeight.w600,
+                  fontFamily: 'LemonMilkMedium'),
+            ),
           ),
+          const SizedBox(height: 5),
+          Text(
+            description,
+            style: TextStyle(
+              color: Colors.white,
+              fontSize: 17.5,
+              fontWeight: FontWeight.w600,
+              fontFamily: 'LeagueSpartan',
+            ),
           ),
         ],
       ),
