+++ lib/core/pages/tracks/tracks_widget.dart
@@ -1,12 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:makeathon7/core/global_variables.dart';
-import 'package:makeathon7/core/pages/tracks/widgets/tracks_page_widget.dart';
+import 'package:makeathon7/core/pages/navigator.dart';
+import 'package:makeathon7/core/pages/tracks/widgets/tracks_description_container.dart';
+import 'package:makeathon7/theme/app_pallete.dart';
 
 class TracksWidget extends StatelessWidget {
   const TracksWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
+    var duration = Duration(milliseconds: 600);
     return SafeArea(
       child: Container(
         decoration: BoxDecoration(
@@ -20,16 +23,67 @@ class TracksWidget extends StatelessWidget {
           appBar: AppBar(
             automaticallyImplyLeading: false,
             backgroundColor: Colors.transparent,
+            leading:
+              IconButton(
+                        onPressed: () {
+                          Navigator.push(
+                              context,
+                              PageRouteBuilder(
+                                pageBuilder:
+                                    (context, animation1, animation2) =>
+                                        HomePageNavigator(),
+                                transitionsBuilder: (context, animation1,
+                                    animation2, child) {
+                                  return FadeTransition(
+                                    opacity: animation1,
+                                    child: child,
+                                  );
+                                },
+                                transitionDuration: duration,
+                              ));
+                        },
+                        icon: Icon(
+                          Icons.arrow_back_ios,
+                          size: MediaQuery.of(context).size.width / 15,
+                          color: AppPallete.whiteColor,
+                        )
+                        ),
           ),
-          body: ListView.builder(
-            itemCount: tracks.length,
-            itemBuilder: (context, index) {
-              final trackDescription = tracks[index];
-              return TracksPageDisplay(
-                title: trackDescription['title'] as String,
-                description: trackDescription['description'] as String,
-              );
-            },
+          body: Center(
+            child: Container(
+              width: 309,
+              height: 625,
+              decoration: BoxDecoration(
+                color: AppPallete.timelineBg,
+                borderRadius: BorderRadius.circular(12),
+                boxShadow: [
+                  BoxShadow(
+                    color: Colors.black.withValues(),
+                    spreadRadius: 4,
+                    blurRadius: 9,
+                    offset: Offset(0, 4), // changes position of shadow
+                  ),
+                ],
+              ),
+              child: Center(
+                child: Column(
+                  children: [
+                    Expanded(
+                      flex: 1,
+                      child: ListView.builder(
+                        itemCount: tracks.length,
+                        itemBuilder: (context, index) {
+                          final trackPage = tracks[index];
+                        return TracksPageDisplay(
+                          title: trackPage['title'] as String,
+                          description: trackPage['description'] as String,
+                        );
+                      }),
+                      ),
+                  ],
+                ),
+              ),
+            ),
           ),
         ),
       ),
