+++ lib/core/pages/tracks/widgets/tracks_button.dart
@@ -17,18 +17,27 @@ class TracksButton extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.all(6.0),
       child: Container(
-        height: 45,
+        height: 47,
         width: 180,
         decoration: BoxDecoration(
-          color: AppPallete.timelineBg,
-                    borderRadius: BorderRadius.circular(27),
-                  ),
+          color: Color.fromARGB(255, 154, 7, 0),
+          borderRadius: BorderRadius.circular(24),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withValues(),
+              spreadRadius: 2.5,
+              blurRadius: 3.5,
+              offset: Offset(0, 4), // changes position of shadow
+            ),
+          ],
+        ),
         child: TextButton.icon(
           onPressed: () {
             Navigator.push(
                 context,
                 PageRouteBuilder(
-                  pageBuilder: (context, animation1, animation2) => TracksWidget(),
+                  pageBuilder: (context, animation1, animation2) =>
+                      TracksWidget(),
                   transitionsBuilder: (context, animation1, animation2, child) {
                     return FadeTransition(
                       opacity: animation1,
@@ -42,7 +51,8 @@ class TracksButton extends StatelessWidget {
               style: TextStyle(
                   color: AppPallete.whiteColor,
                   fontSize: 20,
-                  fontFamily: 'LemonMilkMedium')),
+                  fontWeight: FontWeight.w900,
+                  fontFamily: 'IntroRust')),
           // icon: icon,
         ),
       ),
