+++ pubspec.yaml
@@ -33,7 +33,7 @@ dependencies:
     sdk: flutter
   flutter_popup: ^3.3.4
   firebase_core: ^3.12.0
-  firebase_auth: ^5.5.0
+  firebase_auth: ^5.5.1
   flutter_launcher_icons: ^0.14.3
